+++ src/App.js
@@ -1,56 +1,11 @@
-import React, { useState, useContext } from "react";
+import React, { useState } from "react";
 
 import Header from "./components/Header";
 
-import { UpdateDefaultTime } from "./modals/UpdateDefaultTime";
+import { UpdateDefaultTime } from "./appSections/UpdateDefaultTime";
+import { WeeklyHours } from "./appSections/WeeklyHours";
 
 import { DefaultHrContext } from "./context/DefaultHrContext";
-
-import styled from "@emotion/styled";
-import "./style/WkDay.css";
-
-// styles for list table
-const Rows = styled.li`
-  display: flex;
-  justify-content: space-around;
-  align-items: start;
-  border: solid aliceblue;
-  background-color: white;
-  width: 100%;
-`;
-const RoWData = styled.p`
-  background: #8690e1;
-  width: 100px;
-`;
-const UList = styled.ul`
-  margin: 0;
-  padding: 0;
-  min-width: 100vw;
-`;
-const TimeInput = (props) => {
-  return <span type="time">{props.time}</span>;
-};
-
-const WeeklyHours = (props) => {
-  const daysOfWk = ["Mon", "Tues", "Weds", "Thurs", "Fri"];
-  const { workingHrs } = useContext(DefaultHrContext);
-  const weekLog = daysOfWk.map((day, dayKey) => {
-    return (
-      <Rows key={dayKey}>
-        <RoWData>{day}</RoWData>
-        <RoWData>
-          <TimeInput time={workingHrs.start} />
-        </RoWData>
-        <RoWData>
-          <TimeInput time={workingHrs.quit} />
-        </RoWData>
-        <RoWData>8 hrs</RoWData>
-      </Rows>
-    );
-  });
-
-  return <UList>{weekLog}</UList>;
-};
 
 const App = () => {
   // convert to useReducer
